styles for shop tiles
